--- a/executor/aggfuncs/func_cume_dist.go
+++ b/executor/aggfuncs/func_cume_dist.go
@@ -24,33 +24,35 @@ type cumeDist struct {
 }
 
 type partialResult4CumeDist struct {
-	partialResult4Rank
-	cum int64
+	curIdx   int
+	lastRank int
+	rows     []chunk.Row
 }
 
 func (r *cumeDist) AllocPartialResult() PartialResult {
-	return PartialResult(&partialResult4CumeDist{})
+	return PartialResult(&partialResult4Rank{})
 }
 
 func (r *cumeDist) ResetPartialResult(pr PartialResult) {
-	p := (*partialResult4CumeDist)(pr)
-	p.partialResult4Rank.reset()
-	p.cum = 0
+	p := (*partialResult4Rank)(pr)
+	p.curIdx = 0
+	p.lastRank = 0
+	p.rows = p.rows[:0]
 }
 
 func (r *cumeDist) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) error {
 	p := (*partialResult4CumeDist)(pr)
-	p.partialResult4Rank.updatePartialResult(rowsInGroup, false, r.compareRows)
+	p.rows = append(p.rows, rowsInGroup...)
 	return nil
 }
 
 func (r *cumeDist) AppendFinalResult2Chunk(sctx sessionctx.Context, pr PartialResult, chk *chunk.Chunk) error {
 	p := (*partialResult4CumeDist)(pr)
-	numRows := int64(len(p.results))
-	for p.cum < numRows && p.results[p.cum] == p.results[p.curIdx] {
-		p.cum++
+	numRows := len(p.rows)
+	for p.lastRank < numRows && r.compareRows(p.rows[p.curIdx], p.rows[p.lastRank]) == 0 {
+		p.lastRank++
 	}
 	p.curIdx++
-	chk.AppendFloat64(r.ordinal, float64(p.cum)/float64(numRows))
+	chk.AppendFloat64(r.ordinal, float64(p.lastRank)/float64(numRows))
 	return nil
 }